+++ routes/web.php
@@ -9,7 +9,7 @@ use App\Http\Controllers\AdminController;
 use App\Http\Controllers\Admin\SolicitudEntregaController;
 use App\Http\Controllers\EquipoController;
 use App\Http\Controllers\AsignacionController;
-use App\Http\Controllers\DevolucionController;
+use App\Http\Controllers\Admin\DevolucionController;
 
 
 
@@ -129,11 +129,14 @@ Route::middleware(['auth','role:admin|admin1|admin2'])->group(function () {
     Route::get('/equipos/crear', [EquipoController::class, 'create'])->name('equipos.create');
     Route::post('/equipos', [EquipoController::class, 'store'])->name('equipos.store');
 });
-Route::middleware(['auth', 'role:admin'])->prefix('admin')->name('admin.')->group(function () {
-    // otras rutas admin...
 
-    Route::resource('devoluciones', DevolucionController::class);
-});
 
+Route::middleware(['auth', 'role:admin'])
+    ->prefix('admin/devoluciones')
+    ->name('admin.devoluciones.')
+    ->group(function () {
+        Route::get('/crear', [DevolucionController::class, 'create'])->name('create');
+        Route::post('/', [DevolucionController::class, 'store'])->name('store');
+    });
